hospital/requirement.txt
asgiref==3.8.1
Django==3.2.21
djangorestframework==3.14.0 
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.7.0
pip==24.0
PyJWT==2.10.1
pymongo==4.6.1
pytz==2025.1
sqlparse==0.2.4
